+++ routes/api.php
@@ -262,6 +262,7 @@ Route::group(
         Route::options('request/error-log', 'RequestController@optionsErrorLog');
         Route::get('request/log', 'RequestController@log');
         Route::options('request/log', 'RequestController@optionsLog');
+
         Route::post('request/error-log', 'RequestController@createErrorLog');
     }
 );